remove duplicates from the timezones and improved sorting
